requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
tensorflow-cpu==2.15.0
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
python-multipart==0.0.6
pytest==8.0.0
httpx==0.26.0